Fix issue with lagged-coordinate vectors in CCM algorithm
This is the link to the issue: https://github.com/AndreiHG/CCM_GAH/issues/5#issue-400902309

The fix is as simple as making the difference between the length of the time series, L, and the time variable of the time series. 

This was an issue when trying to do CCM on the mgp93 data because I was passing the data as pandas series instead of as a numpy array, so data[i] would reference to data at index i, instead of the data at position i.

The solution is therefore to pass the data as a numpy array (use .values) and also to select the data based on its position in the data series, not on the index (use .iloc instead of .loc).

I made a comment about this in the jupyter notebook, too.