Pearson coefficient between exactly zero data
When doing CCM between two data series that are both zero, we end up having to compute the Pearson correlation between two zero data sets (the approximated one will also be zero). This causes a division by zero.

While in principle this is an example of perfect matching, and it could be that the two data sets are actually related (say at a later time they both go to some positive value), I think it is best to set the correlation coefficient to zero in these ranges. In general it is not a good idea to draw too many conclusions from sparse data or from a very small relative population.