No duplicates in CCM graph
When using CCM files with more than one embedding dimension E, we also get some duplicates (so pairs of variables that have a large cross-correlation coefficient for more than one embedding dimension). 
We drop the duplicates and keep the pair with the largest CCM cross-correlation coefficient (largest value of the Pearson coefficient). In general, in trying to find the "correct" value of E to use, the value is that for which we get the largest cross-correlation coefficient (see "Advances in nonlinear geosciences", 2018, Sugihara's chapter on CCM).